Refactor PostDetail into reusable components and add paginated replies
- Broke down PostDetail into reusable hooks and components to reduce complexity
- Extracted data-fetching logic into custom hooks (post, community, replies)
- Introduced ReplyTree and ReplyNode components for recursive thread rendering
- Added pagination for root-level comments (8 per page) with reusable Pager component
- Updated comment sorting to prioritize threads with more replies, then oldest to newest
- Hardened reply tree logic to handle non-array inputs safely
- Improved memoization to address React Hook dependency warnings
