+++ src/pages/Community/myCommunity/postDetail/components/PostDetailShell.js
@@ -1,0 +1,13 @@
+const PostDetailShell = ({ heroStyle, PageHeader, Footer, children }) => {
+  return (
+    <section className="ForumContainer">
+      <div className="ForumHero ForumHero--small" style={heroStyle}>
+        <PageHeader />
+      </div>
+      {children}
+      <Footer />
+    </section>
+  );
+};
+
+export default PostDetailShell;
