+++ src/pages/Study/Notes/NotesListDrawer.css
